# internal/docs/swagger.yaml
basePath: /
definitions:
  aclmodels.Acl2Scope:
    enum:
    - ""
    - ror
    - cluster
    - project
    - datacenter
    - virtualmachine
    type: string
    x-enum-comments:
      Acl2ScopeRor: ROR
      Acl2ScopeUnknown: unknown
    x-enum-varnames:
    - Acl2ScopeUnknown
    - Acl2ScopeRor
    - Acl2ScopeCluster
    - Acl2ScopeProject
    - Acl2ScopeDatacenter
    - Acl2ScopeVirtualMachine
...
  apicontracts.AccessControlList:
    properties:
      accessGroups:
        items:
          type: string
        type: array
    type: object
  apicontracts.Agent:
    properties:
      sha:
        type: string
      version:
        type: string
    type: object
...
  apicontracts.Billing:
    properties:
      workorder:
        type: string
    type: object
  apicontracts.Cluster:
    properties:
      acl:
        $ref: '#/definitions/apicontracts.AccessControlList'
      clusterId:
        type: string
      clusterIdOld:
        type: string
      clusterName:
        type: string
      config:
        $ref: '#/definitions/apicontracts.ClusterConfig'
      created:
        type: string
      createdBy:
        type: string
      environment:
        type: string
      firstObserved:
        type: string
      healthStatus:
        $ref: '#/definitions/apicontracts.HealthStatus'
      id:
        type: string
      identifier:
        type: string
      ingresses:
        items:
          $ref: '#/definitions/apicontracts.Ingress'
        type: array
      lastObserved:
        type: string
      metadata:
        $ref: '#/definitions/apicontracts.ClusterMetadata'
      metrics:
        $ref: '#/definitions/apicontracts.Metrics'
      splunkIndex:
        type: string
      status:
        $ref: '#/definitions/apicontracts.ClusterStatus'
      topology:
        $ref: '#/definitions/apicontracts.Topology'
      updated:
        type: string
      versions:
        $ref: '#/definitions/apicontracts.Versions'
      workspace:
        $ref: '#/definitions/apicontracts.Workspace'
      workspaceId:
        type: string
    type: object
  apicontracts.ClusterCondition:
    properties:
      created:
        type: string
      message:
        type: string
      status:
        $ref: '#/definitions/apicontracts.ConditionStatus'
      type:
        $ref: '#/definitions/apicontracts.ConditionType'
      updated:
        type: string
    type: object
  apicontracts.ClusterConfig:
    properties:
      overrides:
        additionalProperties: true
        type: object
      projectMetadata:
        $ref: '#/definitions/apicontracts.ProjectMetadata'
      versions:
        additionalProperties: true
        type: object
    type: object
...
  apicontracts.ClusterMetadata:
    properties:
      billing:
        $ref: '#/definitions/apicontracts.Billing'
      criticality:
        $ref: '#/definitions/apicontracts.CriticalityLevel'
      description:
        minLength: 1
        type: string
      project:
        $ref: '#/definitions/apicontracts.Project'
      projectId:
        type: string
      roles:
        items:
          $ref: '#/definitions/apicontracts.ProjectRole'
        type: array
      sensitivity:
        $ref: '#/definitions/apicontracts.SensitivityLevel'
      serviceTags:
        additionalProperties:
          type: string
        type: object
    type: object
  apicontracts.ClusterPhase:
    enum:
    - Unknown
    - Creating
    - Ready
    - Updating
    - Deleting
    - Deleted
    - Error
    type: string
    x-enum-varnames:
    - ClusterPhaseUnkown
    - ClusterPhaseCreating
    - ClusterPhaseReady
    - ClusterPhaseUpdating
    - ClusterPhaseDeleting
    - ClusterPhaseDeleted
    - ClusterPhaseError
  apicontracts.ClusterState:
    enum:
    - Unknown
    - Ready
    - Warning
    - Error
    type: string
    x-enum-varnames:
    - ClusterStateUnknown
    - ClusterStateReady
    - ClusterStateWarning
    - ClusterStateError
  apicontracts.ClusterStatus:
    properties:
      conditions:
        items:
          $ref: '#/definitions/apicontracts.ClusterCondition'
        type: array
      phase:
        $ref: '#/definitions/apicontracts.ClusterPhase'
      state:
        $ref: '#/definitions/apicontracts.ClusterState'
    type: object
  apicontracts.ConditionStatus:
    enum:
    - Unknown
    - "True"
    - "False"
    type: string
    x-enum-varnames:
    - ConditionStatusUnknown
    - ConditionStatusTrue
    - ConditionStatusFalse
  apicontracts.ConditionType:
    enum:
    - Unknown
    - ClusterReady
    - ToolingOk
    type: string
    x-enum-varnames:
    - ConditionTypeUnknown
    - ConditionTypeClusterReady
    - ConditionTypeToolingOk
  apicontracts.ControlPlane:
    properties:
      metrics:
        $ref: '#/definitions/apicontracts.Metrics'
      nodes:
        items:
          $ref: '#/definitions/apicontracts.Node'
        type: array
    type: object
  apicontracts.CriticalityLevel:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - CriticalityLevelUnknown
    - CriticalityLevelOpen
    - CriticalityLevelIntern
    - CriticalityLevelShielded
    - CriticalityLevelHighlyShielded
  apicontracts.Datacenter:
    properties:
      apiEndpoint:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/apicontracts.DatacenterLocation'
      name:
        type: string
      provider:
        $ref: '#/definitions/providers.ProviderType'
    type: object
  apicontracts.DatacenterLocation:
    properties:
      country:
        type: string
      id:
        type: string
      region:
        type: string
    type: object
...
  apicontracts.EndpointAddress:
    properties:
      nodename:
        type: string
      podnamespace:
        type: string
    type: object
...
  apicontracts.Health:
    enum:
    - 0
    type: integer
    x-enum-varnames:
    - HealthUnknown
  apicontracts.HealthStatus:
    properties:
      health:
        $ref: '#/definitions/apicontracts.Health'
      messages:
        items:
          $ref: '#/definitions/apicontracts.StatusMessage'
        type: array
    type: object
  apicontracts.Ingress:
    properties:
      class:
        type: string
      health:
        $ref: '#/definitions/apicontracts.Health'
      ingressrules:
        items:
          $ref: '#/definitions/apicontracts.IngressRule'
        type: array
      name:
        type: string
      namespace:
        type: string
      uid:
        type: string
    type: object
  apicontracts.IngressPath:
    properties:
      path:
        type: string
      service:
        $ref: '#/definitions/apicontracts.Service'
    type: object
  apicontracts.IngressRule:
    properties:
      hostname:
        type: string
      ipaddresses:
        items:
          type: string
        type: array
      rules:
        items:
          $ref: '#/definitions/apicontracts.IngressPath'
        type: array
    type: object
...
  apicontracts.Metrics:
    properties:
      clusterCount:
        type: integer
      cpu:
        type: integer
      cpuConsumed:
        type: integer
      cpuPercentage:
        type: integer
      memory:
        type: integer
      memoryConsumed:
        type: integer
      memoryPercentage:
        type: integer
      nodeCount:
        type: integer
      nodePoolCount:
        type: integer
      priceMonth:
        type: integer
      priceYear:
        type: integer
    type: object
...
  apicontracts.NhnTooling:
    properties:
      branch:
        type: string
      environment:
        type: string
      version:
        type: string
    type: object
  apicontracts.Node:
    properties:
      architecture:
        type: string
      containerRuntimeVersion:
        type: string
      created:
        type: string
      kernelVersion:
        type: string
      kubeProxyVersion:
        type: string
      kubeletVersion:
        type: string
      machineClass:
        type: string
      machineName:
        type: string
      metrics:
        $ref: '#/definitions/apicontracts.Metrics'
      name:
        type: string
      operatingSystem:
        type: string
      osImage:
        type: string
      role:
        type: string
    type: object
...
  apicontracts.NodePool:
    properties:
      machineClass:
        type: string
      metrics:
        $ref: '#/definitions/apicontracts.Metrics'
      name:
        type: string
      nodes:
        items:
          $ref: '#/definitions/apicontracts.Node'
        type: array
    type: object
...
  apicontracts.PaginatedResult-apicontracts_Cluster:
    properties:
      data:
        items:
          $ref: '#/definitions/apicontracts.Cluster'
        type: array
      dataCount:
        type: integer
      offset:
        type: integer
      totalCount:
        type: integer
    type: object
...
  apicontracts.Project:
    properties:
      active:
        type: boolean
      created:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      projectMetadata:
        $ref: '#/definitions/apicontracts.ProjectMetadata'
      updated:
        type: string
    type: object
  apicontracts.ProjectContactInfo:
    properties:
      email:
        type: string
      phone:
        type: string
      upn:
        type: string
    type: object
  apicontracts.ProjectMetadata:
    properties:
      billing:
        $ref: '#/definitions/apicontracts.Billing'
      roles:
        items:
          $ref: '#/definitions/apicontracts.ProjectRole'
        type: array
      serviceTags:
        additionalProperties:
          type: string
        type: object
    type: object
  apicontracts.ProjectRole:
    properties:
      contactInfo:
        $ref: '#/definitions/apicontracts.ProjectContactInfo'
      roleDefinition:
        $ref: '#/definitions/apicontracts.ProjectRoleDefinition'
    type: object
  apicontracts.ProjectRoleDefinition:
    enum:
    - ""
    - Owner
    - Responsible
    type: string
    x-enum-varnames:
    - ProjectRoleUnknown
    - ProjectRoleOwner
    - ProjectRoleResponsible
  apicontracts.SensitivityLevel:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - SensitivityLevelUnknown
    - SensitivityLevelNormal
    - SensitivityLevelModerate
    - SensitivityLevelHigh
    - SensitivityLevelCritical
  apicontracts.Service:
    properties:
      endpoints:
        items:
          $ref: '#/definitions/apicontracts.EndpointAddress'
        type: array
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/apicontracts.ServicePort'
        type: array
      selector:
        type: string
      type:
        type: string
    type: object
  apicontracts.ServicePort:
    properties:
      name:
        type: string
      nodeport:
        type: string
      protocol:
        type: string
    type: object
...
  apicontracts.StatusMessage:
    properties:
      message:
        type: string
      time:
        type: string
    type: object
...
  apicontracts.Topology:
    properties:
      controlPlane:
        $ref: '#/definitions/apicontracts.ControlPlane'
      controlPlaneEndpoint:
        type: string
      egressIp:
        type: string
      nodePools:
        items:
          $ref: '#/definitions/apicontracts.NodePool'
        minItems: 1
        type: array
    type: object
...
  apicontracts.Versions:
    properties:
      agent:
        $ref: '#/definitions/apicontracts.Agent'
      kubernetes:
        type: string
      nhnTooling:
        $ref: '#/definitions/apicontracts.NhnTooling'
    type: object
  apicontracts.Workspace:
    properties:
      datacenter:
        $ref: '#/definitions/apicontracts.Datacenter'
      datacenterId:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  apiresourcecontracts.CriticalityLevel:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - CriticalityLevelUnknown
    - CriticalityLevelOpen
    - CriticalityLevelIntern
    - CriticalityLevelShielded
    - CriticalityLevelHighlyShielded
  apiresourcecontracts.EnvironmentType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - EnvironmentUnknown
    - EnvironmentDevelopment
    - EnvironmentTesting
    - EnvironmentQA
    - EnvironmentProduction
...
  apiresourcecontracts.ResourceActionType:
    enum:
    - ""
    - Create
    - Update
    - Delete
    type: string
    x-enum-varnames:
    - ResourceActionTypeUnknown
    - ResourceActionTypeCreate
    - ResourceActionTypeUpdate
    - ResourceActionTypeDelete
  apiresourcecontracts.ResourceClusterOrder:
    properties:
      apiVersion:
        type: string
      kind:
        type: string
      metadata:
        $ref: '#/definitions/apiresourcecontracts.ResourceMetadata'
      spec:
        $ref: '#/definitions/apiresourcecontracts.ResourceClusterOrderSpec'
      status:
        $ref: '#/definitions/apiresourcecontracts.ResourceClusterOrderStatus'
    type: object
  apiresourcecontracts.ResourceClusterOrderSpec:
    properties:
      cluster:
        minLength: 1
        type: string
      criticality:
        allOf:
        - $ref: '#/definitions/apiresourcecontracts.CriticalityLevel'
        maximum: 4
        minimum: 1
      environment:
        allOf:
        - $ref: '#/definitions/apiresourcecontracts.EnvironmentType'
        maximum: 4
        minimum: 1
      highAvailability:
        type: boolean
      k8sVersion:
        type: string
      nodePools:
        items:
          $ref: '#/definitions/apiresourcecontracts.ResourceClusterOrderSpecNodePool'
        minItems: 1
        type: array
      orderBy:
        minLength: 1
        type: string
      orderType:
        allOf:
        - $ref: '#/definitions/apiresourcecontracts.ResourceActionType'
        minLength: 1
      ownerGroup:
        minLength: 1
        type: string
      projectId:
        type: string
      provider:
        $ref: '#/definitions/providers.ProviderType'
      providerConfig:
        additionalProperties: true
        type: object
      sensitivity:
        allOf:
        - $ref: '#/definitions/apiresourcecontracts.SensitivityLevel'
        maximum: 4
        minimum: 1
      serviceTags:
        additionalProperties:
          type: string
        type: object
    required:
    - cluster
    - nodePools
    - orderBy
    - orderType
    type: object
  apiresourcecontracts.ResourceClusterOrderSpecNodePool:
    properties:
      count:
        minimum: 1
        type: integer
      machineClass:
        minLength: 1
        type: string
      name:
        minLength: 1
        type: string
    required:
    - count
    - machineClass
    - name
    type: object
  apiresourcecontracts.ResourceClusterOrderStatus:
    properties:
      conditions:
        items:
          $ref: '#/definitions/apiresourcecontracts.ResourceKubernetesClusterStatusCondition'
        type: array
      createdTime:
        type: string
      lastObservedTime:
        type: string
      phase:
        $ref: '#/definitions/apiresourcecontracts.ResourceClusterOrderStatusPhase'
      status:
        type: string
      updatedTime:
        type: string
    type: object
  apiresourcecontracts.ResourceClusterOrderStatusPhase:
    enum:
    - Pending
    - Creating
    - Updating
    - Deleting
    - Completed
    - Failed
    type: string
    x-enum-varnames:
    - ResourceClusterOrderStatusPhasePending
    - ResourceClusterOrderStatusPhaseCreating
    - ResourceClusterOrderStatusPhaseUpdating
    - ResourceClusterOrderStatusPhaseDeleting
    - ResourceClusterOrderStatusPhaseCompleted
    - ResourceClusterOrderStatusPhaseFailed
  apiresourcecontracts.ResourceKubernetesClusterStatusCondition:
    properties:
      lastTransitionTime:
        type: string
      message:
        type: string
      reason:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  apiresourcecontracts.ResourceListClusterorders:
    properties:
      clusterorders:
        items:
          $ref: '#/definitions/apiresourcecontracts.ResourceClusterOrder'
        type: array
      owner:
        $ref: '#/definitions/apiresourcecontracts.ResourceOwnerReference'
    type: object
  apiresourcecontracts.ResourceMetadata:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      creationTimestamp:
        type: string
      generation:
        type: integer
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
      ownerReferences:
        items:
          $ref: '#/definitions/apiresourcecontracts.ResourceMetadataOwnerReference'
        type: array
      resourceVersion:
        type: string
      uid:
        type: string
    type: object
  apiresourcecontracts.ResourceMetadataOwnerReference:
    properties:
      apiVersion:
        type: string
      kind:
        type: string
      name:
        type: string
      uid:
        type: string
    type: object
...
  apiresourcecontracts.ResourceOwnerReference:
    properties:
      scope:
        allOf:
        - $ref: '#/definitions/aclmodels.Acl2Scope'
        description: cluster, workspace,...
      subject:
        description: ror id eg clusterId or workspaceName
        type: string
    type: object
...
  apiresourcecontracts.SensitivityLevel:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - SensitivityLevelUnknown
    - SensitivityLevelNormal
    - SensitivityLevelModerate
    - SensitivityLevelHigh
    - SensitivityLevelCritical
...
  providers.ProviderType:
    enum:
    - unknown
    - tanzu
    - aks
    - k3d
    - kind
    - gke
    - talos
    type: string
    x-enum-varnames:
    - ProviderTypeUnknown
    - ProviderTypeTanzu
    - ProviderTypeAks
    - ProviderTypeK3d
    - ProviderTypeKind
    - ProviderTypeGke
    - ProviderTypeTalos
  rorerror.RorError:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact:
    name: ROR
    url: https://github.com/NorskHelsenett/ror
  description: 'ROR-API, need any help? Go to channel #drift-sdi-devops in norskhelsenett.slack.com
    slack workspace'
  title: Swagger ROR-API
  version: "0.1"
paths:
...
  /v1/orders:
    get:
      consumes:
      - application/json
      description: Orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicontracts.PaginatedResult-apicontracts_Cluster'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rorerror.RorError'
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get orders
      tags:
      - orders
  /v1/orders/{uid}:
    delete:
      consumes:
      - application/json
      description: Orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: bool
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rorerror.RorError'
      security:
      - AccessToken: []
        ApiKey: []
      summary: Delete a order by uid
      tags:
      - orders
    get:
      consumes:
      - application/json
      description: Orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresourcecontracts.ResourceListClusterorders'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rorerror.RorError'
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get order by uid
      tags:
      - orders
  /v1/orders/cluster:
    delete:
      consumes:
      - application/json
      description: Order a kubernetes cluster
      parameters:
      - description: Filter
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/apiresourcecontracts.ResourceClusterOrderSpec'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicontracts.PaginatedResult-apicontracts_Cluster'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rorerror.RorError'
      security:
      - AccessToken: []
        ApiKey: []
      summary: Order deletion a kubernetes cluster
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Order a kubernetes cluster
      parameters:
      - description: Filter
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/apiresourcecontracts.ResourceClusterOrderSpec'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicontracts.PaginatedResult-apicontracts_Cluster'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rorerror.RorError'
      security:
      - AccessToken: []
        ApiKey: []
      summary: Order a kubernetes cluster
      tags:
      - orders
...
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
  ApiKey:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"